requirements.txt
asgiref==3.4.1
boto3==1.18.4
botocore==1.21.4
Django==3.2.5
django-appconf==1.0.4
django-crispy-forms==1.12.0
django-image-cropping==1.5.0
django-imagekit==4.0.2
django-imagekit-cropper==1.17
easy-thumbnails==2.7.1
jmespath==0.10.0
olefile==0.46
pilkit==2.0
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
